Move files to /lib
