+++ src/lib/firmIdCommand.ts
