+++ src/lib/statusBarItem.ts
